+++ src/builder/constants.ts
@@ -1,6 +1,7 @@
 export const APP_NAME = 'appdeveloper';
-export const APP_VERISON = '0.0.4';
+export const APP_VERISON = '0.0.5';
 export const APP_CONFIG_FILE = 'appdev.json';
 export const APP_ARCHITECTURE_DIAGRAM_FILE = 'architecture.md';
 export const CLASS_DIAGRAM_FILE = 'classDiagram.md';
 export const PROJECT_TREE_DIAGRAM_FILE = 'projectTreeDiagram.md';
+export const MAX_RETRY_COUNT = 1;