+++ src/builder/mobile/mobileApp.ts
@@ -1,16 +1,16 @@
-import * as vscode from 'vscode';
 import { App, AppStage, IAppStageInput, IAppStageOutput } from '../app';
 import {
   convertToMermaidMarkdown,
   isMermaidMarkdown,
-  parseResponse,
-  sortComponentsByDependency,
 } from '../utils/contentUtil';
 import { GenerateCodeForComponentPrompt, InitializeAppPrompt } from '../prompt';
 import {
-  IGenerateCodeForComponentResponse,
-  IGenerateCodeResponse,
-  IInitializeAppResponse,
+  ComponetType,
+  ZGenerateCodeForComponentResponseSchema,
+  ZGenerateCodeForComponentResponseType,
+  ZGenerateCodeResponseType,
+  ZInitializeAppResponseSchema,
+  ZInitializeAppResponseType,
 } from '../types';
 import {
   createExpoApp,
@@ -41,14 +41,14 @@ export class MobileApp extends App {
     // Check if node is installed
     const nodeCheck = await checkNodeInstallation();
     if (!nodeCheck.installed) {
-      this.markdown(
+      this.logMessage(
         'Node.js is not installed. Please install Node.js to proceed',
       );
       this.setStage(AppStage.Cancelled);
       return false;
     }
     if (!nodeCheck.meetsMinimum) {
-      this.markdown(
+      this.logMessage(
         `Node.js version ${nodeCheck.version} is not supported. Please install Node.js version 16.0.0 or higher to proceed`,
       );
       this.setStage(AppStage.Cancelled);
@@ -59,81 +59,85 @@ export class MobileApp extends App {
 
   async initialize(
     userMessage?: string,
-  ): Promise<IAppStageOutput<IInitializeAppResponse>> {
+  ): Promise<IAppStageOutput<ZInitializeAppResponseType>> {
     if (!userMessage) {
-      this.markdown(
+      this.logMessage(
         'Please provide a valid input to start building a mobile app',
       );
       this.setStage(AppStage.Cancelled);
       throw new Error('Invalid input');
     }
     this.setStage(AppStage.Initialize);
-    this.markdown('Lets start building a mobile app');
+    this.logMessage('Lets start building a mobile app');
 
     const initializeAppPrompt = new InitializeAppPrompt({
       userMessage: userMessage,
     });
 
     const initializeAppMessages = [
-      vscode.LanguageModelChatMessage.Assistant(MOBILE_BUILDER_INSTRUCTION),
+      this.languageModelService.createSystemMessage(MOBILE_BUILDER_INSTRUCTION),
       // Add user's message
-      vscode.LanguageModelChatMessage.User(initializeAppPrompt.getPromptText()),
+      this.languageModelService.createUserMessage(
+        initializeAppPrompt.getInstructionsPrompt(),
+      ),
     ];
 
     // send the request
-    this.progress('Analyzing app requirements');
-    let createAppResponse, createAppResponseObj;
+    this.logProgress('Analyzing app requirements');
     try {
-      [createAppResponse, createAppResponseObj] =
-        await parseResponse<IInitializeAppResponse>(
-          this.model,
-          initializeAppMessages,
-          this.token,
-          initializeAppPrompt,
+      let { response: createAppResponse, object: createAppResponseObj } =
+        await this.languageModelService.generateObject<ZInitializeAppResponseType>(
+          {
+            messages: initializeAppMessages,
+            schema: ZInitializeAppResponseSchema,
+            responseFormatPrompt: initializeAppPrompt.getResponseFormatPrompt(),
+          },
         );
       initializeAppMessages.push(
-        vscode.LanguageModelChatMessage.Assistant(createAppResponse),
+        this.languageModelService.createAssistantMessage(createAppResponse),
       );
+
+      this.logMessage(`Let's call the app: ${createAppResponseObj.name}`);
+      console.warn(`${JSON.stringify(createAppResponseObj.components)}`);
+      this.logProgress(`Creating app ${createAppResponseObj.name}`);
+      const formattedAppName = createAppResponseObj.name
+        .replace(/\s/g, '-')
+        .toLowerCase();
+      // fix app name
+      createAppResponseObj.name = formattedAppName;
+
+      await this.postInitialize(createAppResponseObj);
+
+      // Create app config
+      const modelConfig = this.languageModelService.getModelConfig();
+      await createAppConfig({
+        name: createAppResponseObj.name,
+        initialPrompt: userMessage,
+        components: JSON.stringify(createAppResponseObj.components),
+        features: createAppResponseObj.features,
+        type: AppType.MOBILE,
+        modelProvider: modelConfig.modelProvider,
+        languageModel: modelConfig.model,
+      });
+
+      return {
+        messages: initializeAppMessages,
+        output: createAppResponseObj,
+      };
     } catch (error) {
       console.error('MobileBuilder: Error parsing response', error);
       throw error;
     }
-
-    this.markdown(`Let's call the app: ${createAppResponseObj.name}`);
-    console.warn(`${JSON.stringify(createAppResponseObj.components)}`);
-
-    this.progress(`Creating app ${createAppResponseObj.name}`);
-    const formattedAppName = createAppResponseObj.name
-      .replace(/\s/g, '-')
-      .toLowerCase();
-    // fix app name
-    createAppResponseObj.name = formattedAppName;
-
-    await this.postInitialize(createAppResponseObj);
-
-    // Create app config
-    await createAppConfig({
-      name: createAppResponseObj.name,
-      initialPrompt: userMessage,
-      components: JSON.stringify(createAppResponseObj.components),
-      features: createAppResponseObj.features,
-      type: AppType.MOBILE,
-    });
-
-    return {
-      messages: initializeAppMessages,
-      output: createAppResponseObj,
-    };
   }
 
-  async postInitialize(createAppResponseObj: IInitializeAppResponse) {
+  async postInitialize(createAppResponseObj: ZInitializeAppResponseType) {
     // Create expo project
     await createExpoApp(createAppResponseObj.name);
     //reset expo project
     await resetExpoProject(createAppResponseObj.name);
-    this.markdown(`Created expo project: ${createAppResponseObj.name}`);
+    this.logMessage(`Created expo project: ${createAppResponseObj.name}`);
     // Design the app
-    this.progress('Writing the design diagram to the file');
+    this.logProgress('Writing the design diagram to the file');
     let designDiagram = createAppResponseObj.design;
     if (!isMermaidMarkdown(designDiagram)) {
       designDiagram = convertToMermaidMarkdown(designDiagram);
@@ -147,34 +151,34 @@ export class MobileApp extends App {
       ],
       createAppResponseObj.name,
     );
-    this.markdown('Design Diagram saved successfully');
+    this.logMessage('Design Diagram saved successfully');
   }
 
   async generateCode({
     previousMessages,
     previousOutput,
-  }: IAppStageInput<IInitializeAppResponse>): Promise<
-    IAppStageOutput<IGenerateCodeResponse>
+  }: IAppStageInput<ZInitializeAppResponseType>): Promise<
+    IAppStageOutput<ZGenerateCodeResponseType>
   > {
     const { name: appName, features, components, design } = previousOutput;
     this.setStage(AppStage.GenerateCode);
 
     // Generate code for each component
-    this.progress('Generating code for components');
+    this.logProgress('Generating code for components');
     // Generate code for individual components first
-    const sortedComponents = sortComponentsByDependency(components);
+    const sortedComponents = this.sortComponentsByDependency(components);
 
     // Generate code for all components
     const generatedCodeByComponent: Map<
       string,
-      IGenerateCodeForComponentResponse
+      ZGenerateCodeForComponentResponseType
     > = new Map();
     let error = false;
     const installedDependencies: string[] = [];
 
     const codeGenerationMessages = [
       ...previousMessages,
-      vscode.LanguageModelChatMessage.User(
+      this.languageModelService.createUserMessage(
         `Lets start generating code for the components one by one.
         Do not create placeholder code.
         Write the actual code that will be used in production.
@@ -199,7 +203,7 @@ export class MobileApp extends App {
       // Generate code for the component
       const codeGenerationPrompt = new GenerateCodeForComponentPrompt({
         name: component.name,
-        type: component.type,
+        type: component.type as ComponetType,
         purpose: component.purpose,
         dependencies: dependenciesWithContent,
         design,
@@ -208,23 +212,27 @@ export class MobileApp extends App {
       });
       const messages = [
         ...codeGenerationMessages,
-        vscode.LanguageModelChatMessage.User(
-          codeGenerationPrompt.getPromptText(),
+        this.languageModelService.createUserMessage(
+          codeGenerationPrompt.getInstructionsPrompt(),
         ),
       ];
 
       let codeGenerationResponse, codeGenerationResponseObj;
       try {
-        this.progress(
+        this.logProgress(
           `Generating code ${componentIndex + 1}/${totalComponents} for component ${component.name}`,
         );
-        [codeGenerationResponse, codeGenerationResponseObj] =
-          await parseResponse<IGenerateCodeForComponentResponse>(
-            this.model,
-            messages,
-            this.token,
-            codeGenerationPrompt,
+        const { response, object } =
+          await this.languageModelService.generateObject<ZGenerateCodeForComponentResponseType>(
+            {
+              messages,
+              schema: ZGenerateCodeForComponentResponseSchema,
+              responseFormatPrompt:
+                codeGenerationPrompt.getResponseFormatPrompt(),
+            },
           );
+        codeGenerationResponse = response;
+        codeGenerationResponseObj = object;
         generatedCodeByComponent.set(component.name, codeGenerationResponseObj);
         // codeGenerationMessages.push(
         //   vscode.LanguageModelChatMessage.Assistant(codeGenerationResponse),
@@ -237,7 +245,7 @@ export class MobileApp extends App {
         ) {
           console.info(`Component ${component.name} has assets`);
           // Save assets
-          this.progress('Saving assets');
+          this.logProgress('Saving assets');
           const files = [];
           for (const asset of codeGenerationResponseObj.assets) {
             files.push({
@@ -246,7 +254,7 @@ export class MobileApp extends App {
             });
           }
           await FileParser.parseAndCreateFiles(files, appName);
-          this.markdown(
+          this.logMessage(
             'Assets saved successfully for component: ' + component.name,
           );
         }
@@ -260,14 +268,16 @@ export class MobileApp extends App {
           component.name,
           error,
         );
-        this.markdown(`Error generating code for component ${component.name}`);
+        this.logMessage(
+          `Error generating code for component ${component.name}`,
+        );
         throw error;
       }
 
-      this.markdown(
+      this.logMessage(
         `Successfully generated code for component ${component.name}`,
       );
-      this.progress(`Writing code to for component ${component.name}`);
+      this.logProgress(`Writing code to for component ${component.name}`);
 
       componentIndex++;
 
@@ -280,13 +290,13 @@ export class MobileApp extends App {
       await FileParser.parseAndCreateFiles(files, appName);
 
       // Install npm dependencies
-      this.progress('Installing npm dependencies');
+      this.logProgress('Installing npm dependencies');
       const npmDependencies = codeGenerationResponseObj.libraries || [];
       installNPMDependencies(appName, npmDependencies, installedDependencies);
 
       // TODO: Check if there are any errors
     }
-    this.progress('Components created successfully');
+    this.logMessage('Components created successfully');
 
     return {
       messages: codeGenerationMessages,